refactor test local rainfall file to use handleResponse function
